P4HW1- use a loop. to calculate an average.
This Program will use a loop. to calculate an average. After displaying score average (after dropping lowest score) , the program is to display a letter grade for the calculated average. This builds on P2HW2 assignment. Instead of an individual statement to collect each score, the program will use a loop. Also, after displaying score average (after dropping lowest score) , the program is to display a letter grade for the calculated average.